refactor(validate-tsconfig-tests): refactor tsconfig tests
this commit refactors & renames
`src/compiler/config/test/validate-ts-sourcemap.spec.ts` to
`src/compiler/config/test/validate-config-sourcemap.spec.ts`

the file was renamed to align with the unit test file
`validate-config.spec.ts`, found in the same directory. This file was
not merged with `validate-config.spec.ts` to avoid more refactoring than
necessary.

the refactoring of the test file consists of:
- renaming the suite & associated tests for clarity
- removing unnecessary filesystem calls
- refactoring generating a `LoadConfigInit` entity out to a helper fn
- variable renaming
